script/pyproject.toml
[project]
name = "script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "aiomysql>=0.2.0",
    "asyncpg>=0.30.0",
    "clickhouse-driver>=0.2.9",
    "cryptography>=46.0.2",
    "motor>=3.7.1",
    "pymongo>=4.15.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
